Create dependencies package, refactor
