+++ src/api/dependencies/access_token.py
@@ -1,0 +1,15 @@
+from fastapi import Depends
+from fastapi_users_db_sqlalchemy import SQLAlchemyAccessTokenDatabase
+from typing import TYPE_CHECKING
+
+from src.models import AccessToken
+from src.core.database.db_helper import db_helper
+
+if TYPE_CHECKING:
+    from sqlalchemy.ext.asyncio import AsyncSession
+
+
+async def get_access_token_db(
+    session: AsyncSession = Depends(db_helper.session_getter),
+):
+    yield SQLAlchemyAccessTokenDatabase(session, AccessToken)
