+++ src/api/dependencies/user.py
@@ -1,0 +1,13 @@
+from fastapi import Depends
+from fastapi_users_db_sqlalchemy import SQLAlchemyUserDatabase
+from typing import TYPE_CHECKING
+
+from src.models import User
+from src.core.database.db_helper import db_helper
+
+if TYPE_CHECKING:
+    from sqlalchemy.ext.asyncio import AsyncSession
+
+
+async def get_user_db(session: AsyncSession = Depends(db_helper.session_getter)):
+    yield SQLAlchemyUserDatabase(session, User)
