+++ src/models/access_token.py
@@ -1,15 +1,11 @@
-from fastapi import Depends
 from fastapi_users_db_sqlalchemy.access_token import (
-    SQLAlchemyAccessTokenDatabase,
     SQLAlchemyBaseAccessTokenTable,
 )
 from sqlalchemy import Integer, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column
-from sqlalchemy.ext.asyncio import AsyncSession
 
 from src.core.types.user_id import UserIDType
 from src.core.database.model import BaseModel
-from src.core.database.db_helper import db_helper
 
 
 class AccessToken(BaseModel, SQLAlchemyBaseAccessTokenTable[UserIDType]):
@@ -18,9 +14,3 @@ class AccessToken(BaseModel, SQLAlchemyBaseAccessTokenTable[UserIDType]):
         ForeignKey("user.id", ondelete="CASCADE"),
         nullable=False,
     )
-
-
-async def get_access_token_db(
-    session: AsyncSession = Depends(db_helper.session_getter),
-):
-    yield SQLAlchemyAccessTokenDatabase(session, AccessToken)
